Moved provisioner client&server into inside provisioner
